feat: add setup command and modal for configuring bot settings
- Implement new `SetupCommand` with chat input command registration
- Add modal functionality to allow users to update AutoHelper channel and monitor channel IDs
- Persist updated settings to the database using `DBManager.editServer`
- Display current and updated settings in the command response
- Add custom IDs for setup button, modal, and text input fields
- Update `ButtonInteractions` to handle the new setup button interaction
- Update type import in `ButtonInteractions.ts`
- Extend `ProcessCache` to support `Command.ChatInputCommandInteraction`
- Update `ValidateFilesCommand` and `Cache` to handle undefined `Processor` in `ProcessCache`
