+++ src/commands/Setup.ts
@@ -3,6 +3,7 @@ import {
   ActionRowBuilder,
   ApplicationIntegrationType,
   ButtonBuilder,
+  ButtonInteraction,
   ButtonStyle,
   ModalActionRowComponentBuilder,
   ModalBuilder,
@@ -15,6 +16,7 @@ import { SetupAhCh, SetupAhMnCh, SetupButton, SetupModal } from '../interaction-
 import { DBManager } from '../Functions/DBManager';
 import { AhChannel } from '../Functions/AutoHelper/AhChannel';
 import { CaseMonitor } from '../Functions/AutoHelper/CaseMonitor';
+import { ProcessCache } from '../CustomTypes/CacheTypes/ProcessCache';
 
 export class SetupCommand extends Command {
   public constructor(context: Command.LoaderContext, options: Command.Options) {
@@ -32,9 +34,8 @@ export class SetupCommand extends Command {
   }
 
   public override async chatInputRun(interaction: Command.ChatInputCommandInteraction) {
-    this.originalInteraction = interaction;
     const serv = Cache.getServer(interaction.guildId!);
-    interaction.reply({
+    const imsg = await interaction.reply({
       embeds: [
         EmbedCreator.Question(
           '__LSPDFR Helper Setup__\n-# Here is some info on your settings!\n\n' +
@@ -50,59 +51,7 @@ export class SetupCommand extends Command {
       ],
       ephemeral: true,
     });
-
-    this.container.client.on(Events.InteractionCreate, async (interaction) => {
-      if (interaction.isButton() && interaction.customId === SetupButton) {
-        await this.originalInteraction?.deleteReply().catch(() => {});
-        const mdl = new ModalBuilder()
-          .setCustomId(SetupModal)
-          .setTitle('Bot Settings')
-          .addComponents([
-            new ActionRowBuilder<ModalActionRowComponentBuilder>().addComponents([
-              new TextInputBuilder()
-                .setCustomId(SetupAhCh)
-                .setLabel('AutoHelper Channel ID')
-                .setStyle(TextInputStyle.Short)
-                .setPlaceholder('0 to disable')
-                .setRequired(true),
-            ]),
-            new ActionRowBuilder<ModalActionRowComponentBuilder>().addComponents([
-              new TextInputBuilder()
-                .setCustomId(SetupAhMnCh)
-                .setLabel('AutoHelper Monitor Channel ID')
-                .setStyle(TextInputStyle.Short)
-                .setPlaceholder('0 to disable')
-                .setRequired(true),
-            ]),
-          ]);
-        await interaction.showModal(mdl);
-      }
-    });
-
-    this.container.client.on(Events.InteractionCreate, async (interaction) => {
-      if (interaction.isModalSubmit() && interaction.customId === SetupModal) {
-        await interaction.reply({ embeds: [EmbedCreator.Loading(`__Processing!__\r\n>>> Changing settings in the DB. Please wait...`)], ephemeral: true });
-        const ahChInput = interaction.fields.getTextInputValue(SetupAhCh);
-        const ahChId = interaction.guild?.channels.cache.get(ahChInput) ? ahChInput : '0';
-        const ahMonChInput = interaction.fields.getTextInputValue(SetupAhMnCh);
-        const ahMonChId = interaction.guild?.channels.cache.get(ahMonChInput) ? ahMonChInput : '0';
-        const servr = Cache.getServer(interaction.guildId!);
-
-        if (!servr) return;
-        servr.ahChId = ahChId;
-        servr.ahMonChId = ahMonChId;
-        await DBManager.editServer(servr);
-        await AhChannel.UpdateCaseMsg(interaction.guildId!);
-        await CaseMonitor.Update(interaction.guildId!);
-        await interaction.editReply({
-          embeds: [
-            EmbedCreator.Question(
-              '__LSPDFR Helper Setup__\n-# Updated your settings!\n\n' +
-                `__**Your New Settings:**__\n**AutoHelper Channel ID:** ${servr?.ahChId}\n**Monitor Channel ID:** ${servr?.ahMonChId}`
-            ),
-          ],
-        });
-      }
-    });
+    const msg = await imsg.fetch();
+    Cache.saveProcess(msg.id, new ProcessCache<undefined>(msg, interaction, undefined));
   }
 }
